storaged: allow dump window to be less than an hour
This is useful for CTS where we just need dump from last few minutes.

Bug: 34198239
Bug: 34845096

+++ storaged/include/storaged.h
@@ -302,7 +302,7 @@ public:
         return mUidm.get_uid_io_stats();
     }
     std::map<uint64_t, std::vector<struct uid_record>> get_uid_records(
-            int hours, uint64_t threshold, bool force_report) {
+            double hours, uint64_t threshold, bool force_report) {
         return mUidm.dump(hours, threshold, force_report);
     }
     void update_uid_io_interval(int interval) {
